transaction_requester: split requester and worker
