+++ gossip/BUILD
@@ -92,6 +92,7 @@ cc_library(
         "//gossip/components:responder_shared",
         "//gossip/components:tips_requester_shared",
         "//gossip/components:transaction_requester_shared",
+        "//gossip/components:transaction_requester_worker_shared",
     ],
 )
 
@@ -107,5 +108,6 @@ cc_library(
         "//gossip/components:responder",
         "//gossip/components:tips_requester",
         "//gossip/components:transaction_requester",
+        "//gossip/components:transaction_requester_worker",
     ],
 )
