+++ gossip/components/BUILD
@@ -64,6 +64,28 @@ cc_library(
 )
 
 cc_library(
+    name = "transaction_requester_worker_shared",
+    hdrs = ["transaction_requester_worker.h"],
+    deps = [
+        ":transaction_requester_shared",
+        "//common:errors",
+    ],
+)
+
+cc_library(
+    name = "transaction_requester_worker",
+    srcs = ["transaction_requester_worker.c"],
+    deps = [
+        ":transaction_requester_worker_shared",
+        "//consensus/tangle",
+        "//gossip:neighbor",
+        "//gossip:node_shared",
+        "//utils:logger_helper",
+        "//utils:time",
+    ],
+)
+
+cc_library(
     name = "transaction_requester_shared",
     hdrs = ["transaction_requester.h"],
     deps = [
